downloader: restructure
